+++ hystrix-dashboard/src/main/java/com/dashboard/HystrixDashboardApplication.java
@@ -1,17 +1,17 @@
-package com.gateway.apigateway;
+package com.dashboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.netflix.hystrix.EnableHystrix;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
 
 @SpringBootApplication
+@EnableHystrixDashboard
 @EnableEurekaClient
-@EnableHystrix
-public class ApiGatewayApplication {
+public class HystrixDashboardApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ApiGatewayApplication.class, args);
+		SpringApplication.run(HystrixDashboardApplication.class, args);
 	}
 
 }